+++ CHANGES.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for `let` expressions.
+
 - Support for tuple expressions of the form `(tuple (name "blockstack") (id 1337))`.
 
 ## 0.1.0 - 2020-09-30