+++ CHANGES.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2020-10-xx
+## [0.2.0] - 2020-10-15
 
 ### Changed
 
@@ -26,4 +26,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The initial public release.
 
-[0.1.1]: https://github.com/weavery/clarity.ml/compare/0.1.0...0.2.0
+[0.2.0]: https://github.com/weavery/clarity.ml/compare/0.1.0...0.2.0